requirements.txt
ammar_filter_post
boto3
botocore
Django
django-cors-headers
djangorestframework
PyJWT
gunicorn
